Permissions subject names are out of order
Problem
-------

An update to OpenDDS caused the order of attributes in the subject
name to become significant.  The stringification of the subject name
does not correspond to the certificate which makes OpenDDS not
recognize the grant.  It seems like BouncyCastle should support this
but no attempts to stringify the subject name produced the correct
order and content.

Solution
--------

Manually stringify the subject for the permissions file.

Signed-off-by: Justin R. Wilson <wilsonj@unityfoundation.io>